Time: 135 ms (44.42%), Space: 108.7 MB (7.02%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -16,7 +16,7 @@ public:
         if(ctr)
             contiguousZeroes.push_back(ctr);
         long long ans = 0;
-        for(long long x: contiguousZeroes)
+        for(long x: contiguousZeroes)
             ans += (x*(x+1))/2;
         return ans;
     }
